fix(1.1.1): Int -> i32

+++ src/transform.rs
@@ -225,7 +225,7 @@ pub(crate) fn convert_field_to_type(field: &Field, import_options: &ImportOption
 
     let scalar = match field_type_name.as_str() {
         "Boolean" => "bool".to_string(),
-        "Int" => "i64".to_string(),
+        "Int" => "i32".to_string(),
         "Float" => "f64".to_string(),
         "String" => "String".to_string(),
         "Json" => "serde_json::Value".to_string(),
